feat(admin): integrate platform admin SPA into monorepo
- Add apps/web/admin to pnpm workspace (Vue 3 + Vite 7 + shadcn-vue)
- Add just targets: admin, admin-build, admin-check, admin-test
- Include admin lint/tests/build in ci-local and full ci pipeline
- Track admin paths in CI workflow changes filter
- Apply CodeRabbit fixes to platform admin backend (waitlist invitation,
  user query, audit repository, controller)

+++ server/smp/src/main/kotlin/com/profiletailors/smp/platformadmin/infrastructure/persistence/R2dbcAdminAuditRepository.kt
@@ -79,7 +79,7 @@ class R2dbcAdminAuditRepository(private val databaseClient: DatabaseClient) :
         }
 
         val where = if (conditions.isEmpty()) "" else "WHERE ${conditions.joinToString(" AND ")}"
-        val offset = query.page * query.size
+        val offset = query.page.toLong() * query.size
 
         val countSql = "SELECT COUNT(*) FROM platform_admin_audit_events $where"
         val dataSql = """
